src/test/java/io/github/vlad324/n775/SolutionTest.java
package io.github.vlad324.n775;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

class SolutionTest {

    private final Solution solution = new Solution();

    private static Stream<Arguments> testDataProvider() {
        return Stream.<Arguments>builder()
            .add(arguments(new int[]{1, 0, 2}, true))
            .add(arguments(new int[]{1, 2, 0}, false))
            .build();
    }

    @ParameterizedTest
    @MethodSource("testDataProvider")
    void should_produce_expected_result(int[] A, boolean expected) {
        // when
        final var actual = solution.isIdealPermutation(A);

        // then
        assertThat(actual).isEqualTo(expected);
    }
}